src/requirements.txt
Flask
flask-socketio
PyJWT
APScheduler
